requirements.txt
streamlit
numpy
pillow
opencv-python-headless
easyocr
torch
ultralytics
paddleocr
paddlepaddle
